@@ -3,47 +3,56 @@
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>@ViewData["Title"] - Chess.Client</title>
-    <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.min.css" />
-    <link rel="stylesheet" href="~/css/site.css" asp-append-version="true" />
-    <link rel="stylesheet" href="~/Chess.Client.styles.css" asp-append-version="true" />
-</head>
-<body>
-    <header>
-        <nav class="navbar navbar-expand-sm navbar-toggleable-sm navbar-light bg-white border-bottom box-shadow mb-3">
-            <div class="container-fluid">
-                <a class="navbar-brand" asp-area="" asp-controller="Home" asp-action="Index">Chess.Client</a>
-                <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target=".navbar-collapse" aria-controls="navbarSupportedContent"
-                        aria-expanded="false" aria-label="Toggle navigation">
-                    <span class="navbar-toggler-icon"></span>
-                </button>
-                <div class="navbar-collapse collapse d-sm-inline-flex justify-content-between">
-                    <ul class="navbar-nav flex-grow-1">
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Index">Home</a>
-                        </li>
-                        <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
-                        </li>
-                    </ul>
-                </div>
-            </div>
-        </nav>
-    </header>
-    <div class="container">
-        <main role="main" class="pb-3">
-            @RenderBody()
-        </main>
-    </div>
+    <title>@ViewData["Title"] - Chess</title>
 
-    <footer class="border-top footer text-muted">
-        <div class="container">
-            &copy; 2022 - Chess.Client - <a asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
-        </div>
-    </footer>
+    <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.min.css" />
+    <link rel="stylesheet" href="~/less/site.min.css" asp-append-version="true" />
+    <link rel="stylesheet" href="~/Chess.Client.styles.css" asp-append-version="true" />
+
+    @* SCRIPTS *@
     <script src="~/lib/jquery/dist/jquery.min.js"></script>
     <script src="~/lib/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
     <script src="~/js/site.js" asp-append-version="true"></script>
+    
+    @* Fontawesome *@
+    <script src="https://kit.fontawesome.com/2e0cc00556.js" crossorigin="anonymous"></script>
+</head>
+
+<body>
+    <header class="header py-2">
+        <div class="container-fluid h-100">
+            <ul class="d-flex justify-content-start align-items-center h-100 list-unstyled">
+                <li class="">
+                    <h1 class="mb-0 fw-normal">
+                        <a asp-controller="Home" asp-action="Index" class="text-decoration-none text-white pointer">
+                            <i class="fa-brands fa-phoenix-framework "></i>
+                        </a>
+                    </h1>
+                </li>
+                
+                <li class="ms-3">
+                    <h4 class="mb-0 fw-normal">
+                        <a asp-controller="Maps" asp-action="Index" class="text-decoration-none text-white">Maps</a>
+                    </h4>
+                </li>
+
+                <li class="align-content-end">ciao</li>
+            </ul>
+        </div>
+    </header>
+
+    <main role="main" class=" py-5">
+        <div class="">
+            @RenderBody()
+        </div>
+    </main>
+
+    @* <footer class="border-top footer text-muted">
+        <div class="container">
+            &copy; 2022 - Chess.Client - <a asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
+        </div>
+    </footer> *@
+
     @await RenderSectionAsync("Scripts", required: false)
 </body>
 </html>
